+++ src/restview/__main__.py
@@ -1,0 +1,7 @@
+"""
+This module contains the entry point to restview.
+"""
+from restview.restviewhttp import main
+
+if __name__ == '__main__':
+    main()